requirements.txt
# HBase AI Doctor Dependencies

# Core framework
# strands - 安装有问题，暂时注释掉
strands-agents
# AWS SDK for Bedrock
boto3>=1.34.0

# Data processing and analysis
pandas>=2.0.0
numpy>=1.24.0

# Date and time handling
python-dateutil>=2.8.0

# Type hints
typing-extensions>=4.0.0

# Logging
# logging - built-in module

# Optional: For better error handling and debugging
rich>=13.0.0